alternator: Use vnode token ranges for shards instead of 1:1 stream ID mapping
CDC stream ids are tied to (scylla) shards per vnode. This means very
many ids per cluster, even with a low number of nodes.

Dynamodb typically expects the number of "shards" in a stream to be much lower.

This patch makes each stream shard a vnode (using the embedded "index" classifier
of cdc stream id), and uses token range query instead of id = <value> (or id in
{ ...values }). This means we can still keep alternator streams stateless (all
info is encoded in the various json identifiers).

There are caveats though: To avoid querying into an old generation, we also
need to make the timestamp restrictions for query adjust according to a shards
creation and ev. expiry time. This is a little problematic for an iterator
with a shards from the latest generation, if a new generation is created
right in the middle of operations.

We assume this is rare, and (mis-)use the local gossip app state for cdc
generation timestamp to check if it seems we expired. If so, we do a
query into cdc tables and get the actual end time to update shard/iterator
in-place. Somewhat expensive, but it should be rare enough.

For old shard iterators, we simply adjust the high threshold for query.
This in turn means that we can include writes made to a new generation
in a new generation query, since we ignore confidence overlap. However,
this is not a problem since we similarly filter out this data from the
new generation query.

Queries to an old generation might be slower though, since they might
hit more than one current vnode. But again, this should be rare, and
adjust itself quickly (catching up).

An added benefit is that we can more easily say that iterators have
hit end-of-data, since end timestamp is now embedded.

Also fixes string length restrictions on ShardId:s. Dynamo define them
as max 65 chars, which we broke even before.

Note that the magic string markers for ShardId and ShardIterator
remain the same, since alternator streams are not GA yet, thus
a format change is ok.

v2:
* Changed asserts in shardid to exceptions as they are parsing user data
* Added asserts in test to verify uniqueness of shards/iterators
* Fixed test assumptions
* Fixed parent shard ordering when doing paged query.
* Fixed shard age threshold for schema
* Fixed token range select bounds
* Added/improved some comments
* Fixed expired shard detection/sequence range in describe

v3:
* Check invalidated shard in get_shard_iterator
* Check expired shard (older than 24h)

v4:
* Removed expired/invalid throws and instead do in-place
  "repair" of shardid/iterators iff expired. This fits
  better with kinesis usage pattern.

v5:
* 24h constant
* More comments and rationale

v6:
* Use lower_bound + -- instead of manual loop.

v6:
* More comments, std::prev, make streams_version pure pod

v7:
* Use cdc::stream tokens as inclusive bounds instead.
* Remove wraparound range workaround and instead handle it properly (split range).
  Now vnode mapping is fully correct.