build: detect availablity of -Wstack-usage=
Detect if the compiler supports -Wstack-usage= and only enable it
if supported. Because each mode uses a different threshold, have each
mode store the threshold value and merge it into the flags only after
we decided we support it.

The switch to make it only a warning is made conditional on compiler
support.